Destroyed main to pieces
