+++ public/static/css/profile.css
@@ -15,6 +15,7 @@
 }
 
 .profile-block > img {
+  margin-top: 20px;
   width: 200px;
   height: 200px;
   border-radius: 100%;
